+++ src/dexterity/localroles/browser/interfaces.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-from zope.interface import Interface
+from zope.interface import Interface, Attribute
 
 
 class IWorkflowState(Interface):
@@ -17,3 +17,17 @@ class IPrincipal(Interface):
 
 class ILocalRoleList(Interface):
     pass
+
+
+class ILocalRoleListUpdatedEvent(Interface):
+    """
+        A LocalRoleList field has been modified
+    """
+
+    fti = Attribute("The Dexterity FTI")
+
+    field = Attribute("The field name that has been changed")
+
+    old_value = Attribute("The value of the field before modification")
+
+    new_value = Attribute("The value of the field after modification")